+++ tiny_dnn/util/target_cost.h
@@ -25,8 +25,8 @@ inline std::vector<size_t> calculate_label_counts(
     }
     label_counts[label]++;
   }
-  assert(std::accumulate(label_counts.begin(), label_counts.end(),
-                         0) == t.size());
+  assert(std::accumulate(label_counts.begin(), label_counts.end(), 0) ==
+         static_cast<int>(t.size()));
   return label_counts;
 }
 
@@ -38,9 +38,7 @@ inline std::vector<size_t> calculate_label_counts(
 //      classes, and the sum of these products equals the total number of
 //      samples)
 inline float_t get_sample_weight_for_balanced_target_cost(
-  size_t classes,
-  size_t total_samples,
-  size_t this_class_samples) {
+  size_t classes, size_t total_samples, size_t this_class_samples) {
   assert(this_class_samples <= total_samples);
   return total_samples / static_cast<float_t>(classes * this_class_samples);
 }
@@ -60,9 +58,9 @@ inline float_t get_sample_weight_for_balanced_target_cost(
 //    * use a value between 0 and 1 to have something between the two extremes
 inline std::vector<vec_t> create_balanced_target_cost(
   const std::vector<label_t> &t, float_t w = 1.0) {
-  const auto label_counts                = calculate_label_counts(t);
+  const auto label_counts         = calculate_label_counts(t);
   const size_t total_sample_count = t.size();
-  const size_t class_count = label_counts.size();
+  const size_t class_count        = label_counts.size();
 
   std::vector<vec_t> target_cost(t.size());
